requirements.txt
requests
Pyexecjs2